requirements.txt
alembic==1.10.4
anyio==3.6.2
asgi-ratelimit==0.10.0
async-timeout==4.0.2
asyncpg==0.27.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
Deprecated==1.2.13
exceptiongroup==1.1.1
fastapi==0.95.1
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
limits==3.4.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
requests==2.29.0
sniffio==1.3.0
SQLAlchemy==1.4.47
SQLAlchemy-Utils==0.41.1
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wrapt==1.15.0
zipp==3.15.0
